requirements.txt
requests>=2.28
PyJWT[crypto]>=2.6
flask>=3.0
